[Fix] `selectors`: unwrap `memo` elements - in both directions.
Note: although this is an important `memo` bugfix, this *might* be a breaking change (note the `find` test change from 1 to 2) but since it should be exceedingly rare for tests to have a reference to both the memoized and unmemoized form of a component, my hope is that it won’t impact anybody.

Fixes #2146.